refactor(overlay): 옵션 네이밍 수정 및 오타 수정, 파일 경로 정리, 훅 확장자 수정
- overlay 옵션 네이밍 변경 (disableInteraction → blockScroll, fullScreen → backdrop)
- OverlayPortal 파일명 오타 수정 및 경로 정리
- useOverlay 훅 확장자 .ts → .tsx 변경
